feat: Bump version to 1.1.19
Fix scan script crash on special characters

The `scan_drive.php` script was crashing when encountering filenames
that contained special characters or non-ASCII characters. This was
caused by the `ExiftoolManager` not specifying the filename character
set.

This commit fixes the issue by adding the `-charset filename=UTF8`
parameter to the `exiftool` command line.

- Modified `scan_drive.php` to add the `-charset` parameter.
- Bumped application version to 1.1.19.
- Updated `CHANGELOG.md` with details of the fix.